requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-yasg
python-decouple
celery
redis
Pillow
psycopg2-binary
pytelegrambotapi